pyproject.toml
[project]
name = "luca"
version = "0.1.0"
description = "Language Transfer AI Tutor - A Pipecat-based voice AI tutor"
requires-python = ">=3.11"

dependencies = [
    "pipecat-ai[daily,deepgram,elevenlabs,cartesia,anthropic,silero]>=0.0.50",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyBKT==1.4.1",
    "rapidfuzz>=3.0.0",
    "unidecode>=1.3.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/luca"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
